pyproject.toml
[project]
name = "cerby-token-retrieval"
authors = [
    { name = "Pedro Santos", email = "pedro@cerby.com" }
]
version = "1.0.0"
description = "Retrieves a bearer token for a Cerby user via local or Okta login non-interactively."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "playwright>=1.56.0",
    "pyotp>=2.9.0",
    "pyyaml>=6.0.3",
]